add check for in-use when deleting a rule

+++ features/editgamerule/src/main/java/host/capitalquiz/editgamerule/domain/GameRuleInteractor.kt
@@ -8,7 +8,7 @@ private const val DEFAULT_GAME_RULE_ID = 1L
 interface GameRuleInteractor {
     fun getAllRules(): Flow<List<GameRule>>
     suspend fun updateRuleForGame(gameId: Long, ruleId: Long)
-    suspend fun deleteRule(ruleId: Long, currentGameId: Long)
+    suspend fun deleteRule(ruleId: Long, currentGameId: Long): Boolean
 
     class Base @Inject constructor(
         private val gameRuleRepository: GameRuleRepository,
@@ -21,12 +21,17 @@ interface GameRuleInteractor {
             gameRuleRepository.setRuleForGame(gameId, ruleId)
         }
 
-        override suspend fun deleteRule(ruleId: Long, currentGameId: Long) {
-            val ruleIdOfCurrentGame = gameRuleRepository.findGameRuleIdOfGame(currentGameId)
-            if (ruleId == ruleIdOfCurrentGame) {
+        override suspend fun deleteRule(ruleId: Long, currentGameId: Long): Boolean {
+            val gamesIdsUsingRule = gameRuleRepository.findAllGamesIdsWithRule(ruleId)
+            val restGamesCount = gamesIdsUsingRule.filterNot { it == currentGameId }.count()
+
+            if (restGamesCount > 0) return false
+
+            if (currentGameId in gamesIdsUsingRule) {
                 gameRuleRepository.setRuleForGame(currentGameId, DEFAULT_GAME_RULE_ID)
             }
             gameRuleRepository.deleteRule(ruleId)
+            return true
         }
     }
 }
